requirements.txt
ttkbootstrap>=1.10
qrcode[svg]>=7.4
Pillow>=10.0
